requirements.txt
amqp==5.3.1
anyio==4.5.2
asgiref==3.8.1
backoff==2.2.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
django==4.2.23
django-celery-beat==2.8.1
django-crontab==0.7.1
django-filter==24.3
django-timezone-field==7.1
exceptiongroup==1.3.0
Faker==35.2.2
flake8==7.1.2
gql==3.5.3
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
kombu==5.5.4
mccabe==0.7.0
multidict==6.1.0
packaging==25.0
promise==2.3
prompt-toolkit==3.0.51
propcache==0.2.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.15.2
